qt, refactor: Move `FreespaceChecker` class into its own module
